python/redshift-cluster/requirements.txt
aws-cdk.core
aws-cdk.aws_redshift
aws_cdk.aws_docdb